Unifies robots (#17)
