+++ pandaReacher/envs/pandaReacherEnv.py
@@ -1,33 +1,25 @@
 import numpy as np
 import pybullet as p
-from abc import abstractmethod
-
-from tiagoReacher.resources.tiagoRobot import TiagoRobot
-from albertReacher.resources.plane import Plane
+from pandaReacher.resources.pandaRobot import PandaRobot
+from pandaReacher.resources.plane import Plane
 from urdfCommon.urdfEnv import UrdfEnv
 
 
-class TiagoReacherEnv(UrdfEnv):
+class PandaReacherEnv(UrdfEnv):
 
-    def __init__(self, render=False, dt=0.01, n=19):
-        super().__init__(TiagoRobot(), render=render, dt=dt)
+    def __init__(self, render=False, dt=0.01, friction=0.0, gripper=False, n=7):
+        super().__init__(PandaRobot(gripper=gripper, friction=friction), render=render, dt=dt)
         self._n = self.robot.n()
         self.setSpaces()
 
     def n(self):
         return self._n
 
-    @abstractmethod
-    def setSpaces(self):
-        pass
-
-    @abstractmethod
-    def applyAction(self, action):
-        pass
-
     def reset(self, initialSet=False, pos=None, vel=None):
-        if not isinstance(pos, np.ndarray) or not pos.size == self._n + 1:
-            pos = np.zeros(self._n+1)
+        if not isinstance(pos, np.ndarray) or not pos.size == self._n:
+            pos = np.zeros(self._n)
+            pos[3] = -1.501
+            pos[5] = 1.8675
         if not isinstance(vel, np.ndarray) or not vel.size == self._n:
             vel = np.zeros(self._n)
         if not initialSet:
