+++ tiagoReacher/envs/vel.py
@@ -2,7 +2,6 @@ from tiagoReacher.envs.tiagoReacherEnv import TiagoReacherEnv
 
 
 class TiagoReacherVelEnv(TiagoReacherEnv):
-    metadata = {"render.modes": ["human"]}
 
     def applyAction(self, action):
         self.robot.apply_base_velocity(action)
